refactor: 패키지구조 [계층/도메인->도메인/계층]으로 변경 (point 도메인)
