--- src/main/java/com/sparta/hhplusconcert/infra/point/PaymentRepository.java
@@ -1,9 +1,0 @@
-package com.sparta.hhplusconcert.infra.point;
-
-import com.sparta.hhplusconcert.domain.point.entity.PaymentEntity;
-
-public interface PaymentRepository {
-
-  Long generatePayment(PaymentEntity payment);
-
-}
