--- src/main/java/com/sparta/hhplusconcert/infra/point/PointHistoryRepository.java
@@ -1,7 +1,0 @@
-package com.sparta.hhplusconcert.infra.point;
-
-import com.sparta.hhplusconcert.domain.point.entity.PointHistoryEntity;
-
-public interface PointHistoryRepository {
-  Long chargePoint(PointHistoryEntity pointHistory);
-}
