+++ src/main/java/com/sparta/hhplusconcert/point/domain/PointTransactionType.java
@@ -1,4 +1,4 @@
-package com.sparta.hhplusconcert.domain.point;
+package com.sparta.hhplusconcert.point.domain;
 
 public enum PointTransactionType {
   CHARGE,
