+++ src/main/java/com/sparta/hhplusconcert/point/domain/User.java
@@ -1,5 +1,6 @@
-package com.sparta.hhplusconcert.api.point.request;
+package com.sparta.hhplusconcert.point.domain;
 
+import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,7 +10,12 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ChargePointRequest {
-  private Long userId;
-  private Long amount;
+public class User {
+
+  private Long id;
+
+  private UUID userUuid;
+
+  private Long point;
+
 }
