+++ src/main/java/com/sparta/hhplusconcert/point/infra/PaymentJpaRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.hhplusconcert.infra.point;
+package com.sparta.hhplusconcert.point.infra;
 
-import com.sparta.hhplusconcert.domain.point.entity.PaymentEntity;
+import com.sparta.hhplusconcert.point.domain.entity.PaymentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentJpaRepository extends JpaRepository<PaymentEntity, Long> {
