+++ src/main/java/com/sparta/hhplusconcert/point/infra/PaymentRepository.java
@@ -1,0 +1,9 @@
+package com.sparta.hhplusconcert.point.infra;
+
+import com.sparta.hhplusconcert.point.domain.entity.PaymentEntity;
+
+public interface PaymentRepository {
+
+  Long generatePayment(PaymentEntity payment);
+
+}
