+++ src/main/java/com/sparta/hhplusconcert/point/infra/PointHistoryJpaRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.hhplusconcert.infra.point;
+package com.sparta.hhplusconcert.point.infra;
 
-import com.sparta.hhplusconcert.domain.point.entity.PointHistoryEntity;
+import com.sparta.hhplusconcert.point.domain.entity.PointHistoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PointHistoryJpaRepository extends JpaRepository<PointHistoryEntity,Long> {
