+++ src/main/java/com/sparta/hhplusconcert/point/infra/PointHistoryRepository.java
@@ -1,0 +1,7 @@
+package com.sparta.hhplusconcert.point.infra;
+
+import com.sparta.hhplusconcert.point.domain.entity.PointHistoryEntity;
+
+public interface PointHistoryRepository {
+  Long chargePoint(PointHistoryEntity pointHistory);
+}
