+++ src/main/java/com/sparta/hhplusconcert/point/infra/PointHistoryRepositoryImpl.java
@@ -1,6 +1,6 @@
-package com.sparta.hhplusconcert.infra.point;
+package com.sparta.hhplusconcert.point.infra;
 
-import com.sparta.hhplusconcert.domain.point.entity.PointHistoryEntity;
+import com.sparta.hhplusconcert.point.domain.entity.PointHistoryEntity;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
