+++ src/main/java/com/sparta/hhplusconcert/point/usecase/ChargePointService.java
@@ -1,10 +1,10 @@
-package com.sparta.hhplusconcert.usecase.point;
+package com.sparta.hhplusconcert.point.usecase;
 
-import com.sparta.hhplusconcert.domain.point.PointTransactionType;
-import com.sparta.hhplusconcert.domain.point.entity.PointHistoryEntity;
-import com.sparta.hhplusconcert.domain.point.entity.UserEntity;
-import com.sparta.hhplusconcert.infra.point.PointHistoryRepositoryImpl;
-import com.sparta.hhplusconcert.infra.point.UserRepositoryImpl;
+import com.sparta.hhplusconcert.point.domain.PointTransactionType;
+import com.sparta.hhplusconcert.point.domain.entity.PointHistoryEntity;
+import com.sparta.hhplusconcert.point.domain.entity.UserEntity;
+import com.sparta.hhplusconcert.point.infra.PointHistoryRepositoryImpl;
+import com.sparta.hhplusconcert.point.infra.UserRepositoryImpl;
 import jakarta.transaction.Transactional;
 import lombok.Builder;
 import lombok.Data;
